+++ src/app/dataset-view/additional-components/data-component/request-timer/request-timer.component.ts
@@ -1,27 +1,81 @@
-import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
+import {
+    ChangeDetectionStrategy,
+    ChangeDetectorRef,
+    Component,
+    Input,
+    OnChanges,
+    OnDestroy,
+    SimpleChanges,
+} from "@angular/core";
 
 @Component({
     selector: "app-request-timer",
     templateUrl: "./request-timer.component.html",
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class RequestTimerComponent {
+export class RequestTimerComponent implements OnChanges, OnDestroy {
+    constructor(private cdr: ChangeDetectorRef) {}
+
+    ngOnDestroy(): void {
+        this.stopTimer();
+    }
+    ngOnChanges(changes: SimpleChanges): void {
+        if (changes.sqlLoading.currentValue) {
+            this.resetTimer();
+            this.runTimer();
+        } else {
+            this.stopTimer();
+        }
+    }
     @Input() public class: string;
-    @Input() public resultTime: number;
+    @Input() public sqlLoading: boolean = false;
+
+    private mm: number = 0;
+    private ss: number = 0;
+    private ms: number = 0;
+    private hh: number = 0;
+    private isRunning: boolean = false;
+    private timerId: NodeJS.Timer;
+
+    private runTimer() {
+        if (!this.isRunning) {
+            this.isRunning = true;
+            this.timerId = setInterval(() => {
+                this.ms++;
+                if (this.ms >= 100) {
+                    this.ss++;
+                    this.ms = 0;
+                }
+                if (this.ss >= 60) {
+                    this.mm++;
+                    this.ss = 0;
+                }
+                if (this.mm >= 60) {
+                    this.hh++;
+                    this.mm = 0;
+                }
+                this.cdr.markForCheck();
+            }, 10);
+        }
+    }
 
     private format(num: number): string {
         return (num + "").length === 1 ? "0" + num : num + "";
     }
 
-    public durationSqlRequest(value: number): string {
-        if (value > 0) {
-            const ms = value % 1000;
-            const ss = Math.floor((value / 1000) % 60);
-            const mm = Math.floor((value / (60 * 1000)) % 60);
-            const hh = Math.floor((value / (3600 * 1000)) % 3600);
-            return `${this.format(hh)}:${this.format(mm)}:${this.format(ss)}.${this.format(ms)}`;
-        } else {
-            return "00:00:00.000";
-        }
+    private resetTimer(): void {
+        this.mm = 0;
+        this.ss = 0;
+        this.ms = 0;
+        this.hh = 0;
+    }
+
+    private stopTimer(): void {
+        this.isRunning = false;
+        clearInterval(this.timerId);
+    }
+
+    public get durationSqlRequest(): string {
+        return `${this.format(this.hh)}:${this.format(this.mm)}:${this.format(this.ss)}.${this.format(this.ms)}`;
     }
 }