+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Added MQTT polling source support 
+- Added MQTT metadata block support
+- Provided 2 mode for compaction: `Full` and `MetadataOnly`
+### Fixed
+- Sql query timer improvements
+
 ## [0.19.0] - 2024-05-14
 ### Added
 - Display dataset ID in Details section(Overview tab)